ideas-api projemize class-validation , error ve loggin özellikleri ek…

+++ src/idea/idea.service.ts
@@ -1,11 +1,13 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus, Logger } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { IdeaEntity } from './idea.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { IdeaDTO } from './idea.dto';
 
+
 @Injectable()
 export class IdeaService {
+  private logger  = new Logger("ideaController");
   constructor(
     @InjectRepository(IdeaEntity)
     private ideaRepository: Repository<IdeaEntity>,
@@ -16,22 +18,36 @@ export class IdeaService {
   }
 
   async ideaById(id: string) {
-    return await this.ideaRepository.findOne({ where: { id } });
+    const idea = await this.ideaRepository.findOne({ where: { id } });
+    if (!idea) {
+      throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+    }
+    return idea;
   }
 
   async create(data: IdeaDTO) {
     const idea = await this.ideaRepository.create(data);
+    this.logger.log(JSON.stringify(data));
     await this.ideaRepository.save(data);
     return idea;
   }
 
   async update(id: string, data: Partial<IdeaDTO>) {
+    const idea = await this.ideaRepository.findOne({ where: { id } });
+    if (!idea) {
+      throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+    }
+    this.logger.log(data);
     await this.ideaRepository.update({ id }, data);
-    return await this.ideaRepository.findOne({ id });
+    return idea;
   }
 
   async delete(id: string) {
+    const idea = await this.ideaRepository.findOne({ where: { id } });
+    if (!idea) {
+      throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+    }
     await this.ideaRepository.delete({ id });
-    return { deleted: true };
+    return idea;
   }
 }
